app/lib/pyproject.toml
[build-system]
requires = ["hatchling", "wheel"]
build-backend = "hatchling.build"

[project]
name = "wyolo"
version = "1.0.0"
description = "Professional YOLO training library with MLOps integration"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "William Steve Rodriguez Villamizar", email = "wisrovi.rodriguez@gmail.com"}
]
maintainers = [
    {name = "William Steve Rodriguez Villamizar", email = "wisrovi.rodriguez@gmail.com"}
]
keywords = ["yolo", "computer-vision", "machine-learning", "mlflow", "mlops"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.8"
dependencies = [
    "click>=8.0.0",
    "dvc>=2.0.0",
    "GPUtil>=1.4.0",
    "loguru>=0.6.0",
    "mlflow>=2.0.0",
    "pillow>=9.0.0",
    "pyyaml>=6.0",
    "python-slugify>=6.0.0",
    "redis>=4.0.0",
    "ultralytics>=8.0.0",
    "wredis>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "black>=22.0.0",
    "bump2version>=1.0.0",
    "build>=0.8.0",
    "flake8>=5.0.0",
    "isort>=5.10.0",
    "mypy>=1.0.0",
    "pylint>=2.17.0",
    "pre-commit>=2.20.0",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.8.0",
    "twine>=4.0.0",
    "types-PyYAML",
    "types-Pillow",
]
docs = [
    "mkdocs>=1.4.0",
    "mkdocs-material>=8.5.0",
    "mkdocstrings[python]>=0.19.0",
]
tune = [
    "ray[tune]>=2.0.0",
]

[project.urls]
Homepage = "https://github.com/wisrovi/wyoloservice2_worker"
Documentation = "https://wyolo.readthedocs.io/"
Repository = "https://github.com/wisrovi/wyoloservice2_worker.git"
"Bug Tracker" = "https://github.com/wisrovi/wyoloservice2_worker/issues"

[project.scripts]
wyolo-train = "wyolo.core.yolo_train:train"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
wyolo = ["py.typed"]

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["wyolo"]
known_third_party = [
    "ultralytics", "mlflow", "loguru", "click", "yaml", 
    "PIL", "GPUtil", "pandas", "matplotlib", "seaborn", 
    "dvc", "wredis", "slugify", "redis"
]
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "ultralytics.*",
    "mlflow.*", 
    "GPUtil.*",
    "dvc.*",
    "wredis.*",
    "slugify.*"
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
]

[tool.pylint.messages_control]
disable = [
    "R0903",  # Too few public methods
    "R0913",  # Too many arguments
    "W0613",  # Unused argument
    "W0622",  # Redefining built-in
    "C0114",  # Missing module docstring (handled separately)
    "C0115",  # Missing class docstring (handled separately)
    "C0116",  # Missing function docstring (handled separately)
]

[tool.pylint.format]
max-line-length = 88

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]